requirements.txt
# Core dependencies
playwright==1.58.0
python-dotenv==1.2.2

# Web server (Electron backend)
flask==2.3.3
flask-cors==6.0.2

# Excel handling
pandas==3.0.1
openpyxl==3.1.5

# 2FA code generation
pyotp==2.9.0

# HTTP
requests==2.28.2

# Utilities
python-dateutil==2.9.0.post0
colorama==0.4.6

# Logging and monitoring
loguru==0.7.0

# Google Drive upload (OAuth2)
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
